refactor: rm userThunks
