+++ src/store/slices/users/usersSlice.ts
@@ -41,5 +41,4 @@ const usersSlice = createSlice({
   },
 });
 
-// export const {  } = usersSlice.actions;
 export default usersSlice.reducer;
